backend-fastapi/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
openai>=1.3.0
firebase-admin>=6.2.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
httpx>=0.24.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
aiofiles>=23.1.0
jinja2>=3.1.2
markdown>=3.4.3
PyPDF2>=3.0.1
pillow>=9.5.0
numpy>=1.24.3
pandas>=2.0.0
matplotlib>=3.7.1
docker>=6.1.2